new kNN implementation
Reworked original kNN implementation to look more like outline provided and be more readable. It did run a good 1.5 minutes slower on the medium dataset, so I implemented a multiprocessing option to take it from around 260 seconds to 15 seconds on the medium dataset. However, windows antivirus seems to prevent multiprocessing on IDLE, so I made a jupyter notebook version, but both are exactly the same.